Implement the base class for the algorithm that inherit the `CharIndexCaretPositionAlgorithm`. Also, move the test case into the base class because all the test case is testing the logic inside the `CharIndexCaretPositionAlgorithm`.
